+++ docs/Changelog.md
@@ -8,6 +8,7 @@
 
 ### `3.8.0`
 
+ * Connections container class (@sirn-se)
  * Dispach onDisconnect on remote connection close (@srebb, @sirn-se)
  * Exceptions: ControlInterface on CloseException, ReconnectException (@sirn-se)
  * Exceptions: HandlerLevelInterface on BadUriException, ClientException, RunnerException, ServerException (@sirn-se)